+++ FolderView.m
@@ -228,7 +228,7 @@
 	{
 		NSMutableArray * array = [NSMutableArray arrayWithCapacity:[self numberOfSelectedRows]];
 		NSIndexSet * selectedRowIndexes = [self selectedRowIndexes];
-		unsigned int item = [selectedRowIndexes firstIndex];
+		NSUInteger item = [selectedRowIndexes firstIndex];
 		
 		while (item != NSNotFound)
 		{
@@ -284,7 +284,7 @@
 -(void)highlightSelectionInClipRect:(NSRect)rect
 {
 	NSIndexSet * selectedRowIndexes = [self selectedRowIndexes];
-	unsigned int rowIndex = [selectedRowIndexes firstIndex];
+	NSUInteger rowIndex = [selectedRowIndexes firstIndex];
 
 	while (rowIndex != NSNotFound)
 	{
